kestra/requirements_kestra.txt
requests
kafka-python
geopandas
pyogrio

